Add new DashboardCategory and MealPage views
Introduced new views to handle category-based meal selection: DashboardCategory and MealPage. DashboardCategory is a custom control used to display meal categories on the dashboard which when clicked will redirect to the MealPage. The MealPage displays specific meals for the selected category. Updated the layout of the dashboard and added different meal icon resources.