Add Xamarin.CommunityToolkit and tap gesture recognizer
The Xamarin.CommunityToolkit package has been added to the .csproj files for the SeniorProjectHealthApplication, Android, and iOS. In the DashboardCategory view, a tap gesture recognizer has been added within a Grid. The gesture is handled by the method OnTapGestureTapped in the DashboardCategory view's code-behind. This results in a cleaner and more maintainable way of capturing and handling the tap event.