Update Xamarin package versions and add Xamarin.CommunityToolkit
This commit updates the versions of Xamarin.Forms and Xamarin.Essentials in the SeniorProjectHealthApplication. Additionally, it includes the Xamarin.CommunityToolkit package across all the application projects (main, Android, and iOS). This would maintain consistency in Xamarin package versions across the different platforms and provide access to additional features and helpers from the toolkit.