Add AddFoodPage with back and barcode functionality
Created a new AddFoodPage in the Food view and implemented functionality for a back button and a barcode button. This page was originally named MealPage and has been renamed to AddFoodPage. Furthermore, unnecessary using directives in multiple database models were removed and the syntax of primary key declarations in the same models was improved.