Update food item data and include OpenFoodFacts package
In this commit, the fetching parameters in FoodCategoryViewModel have been updated for more accurate results. Also, the OpenFoodFacts-CSharp.csproj package is newly referenced in the project. Now, it fetches data with a different product code ("01289903" instead of "034856890089") and searches for "chicken" rather than "raw chicken breast". This could give better search results for users looking for food items or recipes.