Add search functionality and remove food item feature
Implemented a search functionality in the AddFoodPage ViewModel by creating a new SearchCommand. Designed the behavior and binding for the search entry in the XAML page. Moreover, added a feature to remove a specific food item from the list. The remove button click event was handled accordingly to execute the newly defined RemoveFoodItem method in FoodCategory ViewModel, to manage database interactions for the deletion operation.