Update search functionality and refactor bindings in AddFood layout
The update includes changes in event handling from "TextChanged" to "Completed" in the SearchCommand. Also, 'FoodName' and 'Calories' bindings were refactored to 'Food_Name' and 'Total_Calories' respectively. New code to search for foods by name was added to the AddFoodPageViewModel. The layout for displaying the search results was also improved with the implementation of a CollectionView. A new FoodDatabaseManager class was also created to manage food data requests.