Update navigation and interactivity across application
The changes focus on improving navigation and interactivity in different areas of the application, namely the DashboardPage, BrowseRecipePage, RecipesPage, and WorkoutsMainPage. The respective .xaml and .cs files have been updated to enhance UI responsiveness and alignment. Additionally, an AddRecipeItem function was added in RecipesViewModel to boost the usability of the application.